refactored and product variants spec boot
