+++ package.json
@@ -38,9 +38,9 @@
   },
   "dependencies": {
     "@crystallize/content-transformer": "^9.1.2",
-    "@crystallize/import-export-sdk": "^0.0.1",
+    "@crystallize/import-export-sdk": "/Users/simakovic/code/crystallize/libraries/components/import-export-sdk",
     "@crystallize/js-api-client": "^1.3.1",
-    "@crystallize/schema": "^0.0.4",
+    "@crystallize/schema": "/Users/simakovic/code/crystallize/libraries/components/schema",
     "dotenv": "^10.0.0",
     "download": "^8.0.0",
     "duration": "^0.2.2",
