+++ src/bootstrap-tenant/json-spec/core.ts
@@ -9,7 +9,7 @@ import {
   Shape,
   ShapeComponent,
   ShapeComponentConfig,
-} from '@crystallize/schema/shape'
+} from '@crystallize/schema'
 import {
   SubscriptionPeriodUnit,
   SubscriptionPlanPeriod,
@@ -169,15 +169,23 @@ export interface JSONPropertiesTableSection {
   properties?: Record<string, any>
 }
 export type JSONPropertiesTable = JSONPropertiesTableSection[]
-export interface JSONItemReference {
+export type JSONItemReference = {
   cataloguePath?: string
   externalReference?: string
+  sku?: string
 }
 export interface JSONGridReference {
   name?: string
 }
-export type JSONItemRelation = JSONItemReference
-export type JSONItemRelations = JSONItemRelation[]
+// export type JSONItemRelation = {
+//   items: JSONItemReference[]
+//   productVariants: JSONItemReference[]
+// }
+// export type JSONItemRelations = [JSONItemRelation]
+export type JSONItemRelations = {
+  items: JSONItemReference[]
+  productVariants: JSONItemReference[]
+}
 export type JSONGridRelation = JSONGridReference
 export type JSONGridRelations = JSONGridReference[]
