+++ src/graphql/build-get-items-by-shape-query.ts
@@ -33,7 +33,7 @@ const basicComponentContent = [
     items: {
       id: true,
     },
-    productVariations: {
+    productVariants: {
       sku: true,
     },
   },
