+++ src/types/shapes/components/item-relations/item-relations-content.input.ts
@@ -2,7 +2,7 @@ import { ComponentContentInput } from '../component-content.input'
 
 export interface ItemRelationsContentInput {
   itemIds: string[]
-  productVariationSkus: string[]
+  skus: string[]
 }
 
 export interface ItemRelationsComponentContentInput
